# Make FxHash the default HashMap hasher