--- a/crates/bevy_ecs/src/schedule_v3/graph_utils.rs
+++ b/crates/bevy_ecs/src/schedule_v3/graph_utils.rs
@@ -2,7 +2,7 @@ use std::fmt::Debug;
 
 use bevy_utils::{
     petgraph::{graphmap::NodeTrait, prelude::*},
-    HashMap, HashSet, FixedState,
+    FixedState, HashMap, HashSet,
 };
 use fixedbitset::FixedBitSet;