# Objective
Speed up all `HashMap` queries in the engine. Shrink the dependency tree.

## Solution
Instead of using `aHash` as the default hasher for all of the engine's HashMap and HashSets, use `FxHash` instead. This collapses `StableHashMap` and `StableHashSet` into the normal `HashMap`/`HashSet`.

This also removes the `fxhash` dependency on `bevy_ecs`, instead having `bevy_utils` depend on `rustc-hash` instead, which is better maintained, hit 1.0 already, and has zero dependencies.

This does mean that none of the types should now be exposed to external untrusted inputs (i.e. networking), since FxHash is not HashDOS resistant.

## Performance
From an [external source](https://nnethercote.github.io/perf-book/hashing.html), performance could increase as much as 4% in HashMap bound code.

---

## Changelog
Removed: `bevy::utils::AHash`.
Removed: `bevy::utils::StableHashMap`.
Removed: `bevy::utils::StableHashSet`.

## Migration Guide
If you were relying on `bevy::utils::StableHashMap` or `bevy::utils::StableHashSet`, use `bevy::utils::HashMap` and `bevy::utils::HashSet` instead. The default hash map and set now have a stable iteration order.