@james7132 2023-01-06
From a gut instinct, the only cases where we weren't using FxHash already is in reflection and assets, and by proxy rendering. The only major use in a hot location in ECS now is just `Components`, which already uses FxHash.

Rendering, in particular, I don't expect large changes since the bottleneck lies elsewhere.