@AxiomaticSemantics 2023-01-06
> From a gut instinct, the only cases where we weren't using FxHash already is in reflection and assets, and by proxy rendering. The only major use in a hot location in ECS now is just `Components`, which already uses FxHash.
> 
> Rendering, in particular, I don't expect large changes since the bottleneck lies elsewhere.
I've noticed no change at all in code that is mutating 1_000_000 entities in 500 systems, if that means anything.