@AxiomaticSemantics 2023-01-06
> In the early days of Bevy I chose aHash over FXHash because it performed measurably better for small index-sized keys (this was back when we were using `HashMap<Entity, EntityLocation>` with "stable entity uuids") The scope of the benchmarks was limited to bevy_ecs. Now that this is piped through "everywhere", I'm curious to see what the perf difference is.

I've noticed no change at all in code that is mutating 1_000_000 entities in 500 systems, if that means anything.