@recatek 2023-01-06
Adding some context from the discord discussion the other day around this: Gut check from reading the related material is that FxHash is faster for the act of hashing on small keys (not necessarily larger keys), whereas aHash is more likely to aid in avoiding collisions in storage and lookup. The difference between the two is minor and it's difficult to determine the best without profiling. Consolidating dependencies would also be good in its own right.

If we _do_ end up doing benchmark comparisons on this, might be worth throwing xxhash-rust into the mix.