@alice-i-cecile 2023-01-08
I'm happy to merge this without further benchmarking based on @AxiomaticSemantics's experiments once CI is passing.