@james7132 2023-01-09
I actually want to fully test this on some stress tests outside of pure ECS tests, especially for rendering, since the only real API heavily reliant on HashMaps in the ECS is World/EntityRef::get and friends, and those are already using FxHash. I just haven't had the time to run them.