@james7132 2023-01-17
Finally had some time to dig into this. Compared this against `main` with the typical `many_foxes` stress test with the `stress-test` profile.

On a frame level, there seems to be a ~2.9% (4.08ms -> 3.96ms) performance improvement.

![image](https://user-images.githubusercontent.com/3137680/212827490-686319ab-6c59-494a-a212-5d3daeadf478.png)

When looking at `main_opaque_pass_3d`, the render graph node in particular that is fairly heavily reliant on faster HashMap lookups, there is a similar 2% improvement (341.64us -> 334.54us)

![image](https://user-images.githubusercontent.com/3137680/212827880-926bb6a4-d8c3-4bc4-963f-ef480dea84ef.png)

`queue_material_meshes`, which is also reliant on the same hashmap lookups, improved ~7% (84.93us -> 79.55%).

These results are in line with the expectations in the aforementioned performance guidance about hashing (single digit percentage improvements in use cases bound by HashMap lookup.

With that said, this is still focused primarily on a case where there's really only one entry in the HashMap, so it's probably not benefiting from the faster hashing, but it's unclear if the quality of it's output will have a more pronounced impact on more varied scenes.