@james7132 2023-01-18
Holding off from merging this since:

 * Recent tests against main show no reasonable difference, and the lack of the default BuildHasher impl from hashbrown is a usability regression.
  * We cannot remove ahash from our dependency tree as `gpu-descriptor` is reliant on it.

As it currently stands, that's probably not enough to justify merging this.