@alice-i-cecile 2023-01-18
Agreed. I'm going to close this out for now. Dear reader, feel free to revive this PR if you have a good argument for why it should be merged.